+++ Realm/RLMArray_Private.hpp
@@ -82,10 +82,12 @@
 // TableView backed RLMArray subclass
 //
 @interface RLMArrayTableView : RLMArray {
+    tightdb::Query _backingQuery;
     tightdb::TableView _backingView;
+    BOOL _viewCreated;
 }
 + (instancetype)arrayWithObjectClassName:(NSString *)objectClassName
-                                    view:(tightdb::TableView const &)view
+                                   query:(tightdb::Query &)query
                                    realm:(RLMRealm *)realm;
 
 @end
@@ -96,7 +98,7 @@
 //
 @interface RLMCArrayHolder : NSObject {
 @public
-    __strong id *array;
+    std::unique_ptr<id[]> array;
     NSUInteger size;
 }